+++ packages/apps/plugins/plugin-chain/src/components/ChainMain.tsx
@@ -1,80 +1,27 @@
 //
-// Copyright 2023 DXOS.org
+// Copyright 2024 DXOS.org
 //
 
 import React, { type FC } from 'react';
 
 import { type ChainType } from '@braneframe/types';
-import { getSpace } from '@dxos/react-client/echo';
-import { DensityProvider, Main, Select, useTranslation } from '@dxos/react-ui';
+import { Main } from '@dxos/react-ui';
 import {
   baseSurface,
-  bottombarBlockPaddingEnd,
-  mx,
-  textBlockWidth,
   topbarBlockPaddingStart,
+  bottombarBlockPaddingEnd,
+  textBlockWidth,
+  mx,
 } from '@dxos/react-ui-theme';
-import { nonNullable } from '@dxos/util';
 
-import { PromptTemplate, Section } from './PromptTemplate';
-import { CHAIN_PLUGIN } from '../meta';
-import { type Preset, presets } from '../presets';
+import { Chain } from './Chain';
 
-const ChainMain: FC<{ chain: ChainType }> = ({ chain }) => {
-  const space = getSpace(chain);
-  if (!space) {
-    return null;
-  }
-
-  const handleSelectPreset = (preset: Preset) => {
-    chain.title = preset.title;
-    chain.prompts.filter(nonNullable).forEach((p) => space.db.remove(p));
-    chain.prompts = [preset.prompt()];
-  };
-
-  return (
-    <Main.Content classNames={[baseSurface, topbarBlockPaddingStart, bottombarBlockPaddingEnd]}>
-      <div role='none' className={mx(textBlockWidth, 'pli-2')}>
-        <div className='flex flex-col my-2'>
-          {chain.prompts.filter(nonNullable).map((prompt, i) => (
-            <PromptTemplate key={i} prompt={prompt} />
-          ))}
-        </div>
-        <Presets presets={presets} onSelect={handleSelectPreset} />
-      </div>
-    </Main.Content>
-  );
-};
-
-const Presets: FC<{ presets: Preset[]; onSelect: (preset: Preset) => void }> = ({ presets, onSelect }) => {
-  const { t } = useTranslation(CHAIN_PLUGIN);
-
-  return (
-    <DensityProvider density='fine'>
-      <Section title='Presets'>
-        <div className='p-2'>
-          <Select.Root
-            onValueChange={(value) => {
-              onSelect(presets.find(({ id }) => id === value)!);
-            }}
-          >
-            <Select.TriggerButton placeholder={t('select preset template placeholder')} />
-            <Select.Portal>
-              <Select.Content>
-                <Select.Viewport>
-                  {presets.map(({ id, title }) => (
-                    <Select.Option key={id} value={id}>
-                      {title}
-                    </Select.Option>
-                  ))}
-                </Select.Viewport>
-              </Select.Content>
-            </Select.Portal>
-          </Select.Root>
-        </div>
-      </Section>
-    </DensityProvider>
-  );
-};
+const ChainMain: FC<{ chain: ChainType }> = ({ chain }) => (
+  <Main.Content classNames={[baseSurface, topbarBlockPaddingStart, bottombarBlockPaddingEnd, textBlockWidth]}>
+    <div role='none' className={mx(textBlockWidth, 'pli-2')}>
+      <Chain chain={chain} />
+    </div>
+  </Main.Content>
+);
 
 export default ChainMain;
