+++ packages/apps/plugins/plugin-chain/src/components/index.ts
@@ -6,3 +6,4 @@ import React from 'react';
 
 // Lazily load components for content surfaces.
 export const ChainMain = React.lazy(() => import('./ChainMain'));
+export const ChainArticle = React.lazy(() => import('./ChainArticle'));
