+++ packages/apps/plugins/plugin-table/src/components/TableArticle.tsx
@@ -1,5 +1,5 @@
 //
-// Copyright 2023 DXOS.org
+// Copyright 2024 DXOS.org
 //
 
 import React, { type FC } from 'react';
